requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
openai==1.30.1